refactor(consts): Move PAGE_SIZE into arch
