+++ src/arch/x86_64/paging/mod.rs
@@ -7,10 +7,11 @@ use x86_64::{
 	},
 };
 
+use super::PAGE_SIZE;
 use crate::{
 	consts::{
-		BOOT_GDT_CODE, BOOT_GDT_DATA, BOOT_GDT_NULL, GDT_OFFSET, KERNEL_OFFSET, PAGE_SIZE,
-		PAGETABLES_END, PAGETABLES_OFFSET, PML4_OFFSET,
+		BOOT_GDT_CODE, BOOT_GDT_DATA, BOOT_GDT_NULL, GDT_OFFSET, KERNEL_OFFSET, PAGETABLES_END,
+		PAGETABLES_OFFSET, PML4_OFFSET,
 	},
 	paging::BumpAllocator,
 };
